project/backend/requirements.txt
asgiref==3.8.1
coverage==7.6.12
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-test-plus==2.2.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
factory_boy==3.3.3
Faker==36.1.1
iniconfig==2.0.0
Markdown==3.7
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
sqlparse==0.5.3
tzdata==2025.1
